requirements.txt
certifi==2018.8.24
chardet==3.0.4
Click==7.0
cloudpickle==0.5.6
cycler==0.10.0
dask==0.19.2
decorator==4.3.0
graphviz==0.8.1
idna==2.6
kiwisolver==1.0.1
matplotlib==3.0.0
networkx==2.2
numpy==1.13.3
Pillow==5.3.0
pyparsing==2.2.2
python-dateutil==2.7.3
PyWavelets==1.0.1
requests==2.18.4
scikit-image==0.14.1
scipy==1.1.0
urllib3==1.22
six==1.11.0
toolz==0.9.0
urllib3==1.22
